+++ ampligraph/latent_features/models/__init__.py
@@ -4,5 +4,6 @@ from .DistMult import DistMult
 from .ComplEx import ComplEx
 from .HolE import HolE
 from .RandomBaseline import RandomBaseline
+from .ConvKB import ConvKB
 
-__all__ = ['EmbeddingModel', 'TransE', 'DistMult', 'ComplEx', 'HolE', 'RandomBaseline']
+__all__ = ['EmbeddingModel', 'TransE', 'DistMult', 'ComplEx', 'HolE', 'ConvKB', 'RandomBaseline']